update: move PictureCache to another file
