+++ PicturePlayground/PicturePlayground/PicturePlaygroundApp.swift
@@ -1,0 +1,17 @@
+//
+//  PicturePlaygroundApp.swift
+//  PicturePlayground
+//
+//  Created by Ahmed Shendy on 09/10/2023.
+//
+
+import SwiftUI
+
+@main
+struct PicturePlaygroundApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
